assignments: Introduce mini-libc assignment
Mini-libc is a minimalistic standard C library (libc) implementation. It
is meant to be used as a complete replacement for the system libc (glibc
in Linux). This is why test programs are built with `-nostinc`,
`-nostlib`, `-fno-builtin` options.

The actual library source code is part of the `libc/` directory. It is
to be compiled as a static library in the `libc.a` file.

A set of sample use cases, as C source files, are stored in the
`samples/` directory.

Tests are in the `tests/` directory, both as source code and shell
scripts.

The aim is to have a minimally functional libc with features such as:

- string management
- very basic memory support
- POSIX file I/O (i.e. `open()`, `read()`, `write()`)
- basic printing (`puts()`)
- sleep functions

Co-authored-by: Ruxandra Grigorie <ruxi.grigorie@gmail.com>
Signed-off-by: Razvan Deaconescu <razvan.deaconescu@upb.ro>